+++ CHANGELOG.md
@@ -7,8 +7,10 @@
 ### 🐞 Bug fixes
 - _...Add new stuff here..._
 
-## 5.6.0-pre.3
+## 5.6.0-pre.4
 ### ✨ Features and improvements
+- Update Maplibre-Native to v6.4.0
+- Remove Hillshade and Color relief workaround since it is now supported in maplibre-native ([#2044](https://github.com/maptiler/tileserver-gl/pull/2044)) (by [acalcutt](https://github.com/acalcutt))
 - feat: Add ignore-missing-files cli option to avoid crashing at startup ([#1896](https://github.com/maptiler/tileserver-gl/pull/1896)) (by [andrewlaguna824](https://github.com/andrewlaguna824))
 
 ### 🐞 Bug fixes